Update pang type from float to double in Player class
Changed the type of 'pang' from float to double in the Player class.

+++ src/com/yusuf/ballwars/Player.java
@@ -2,11 +2,11 @@ package com.yusuf.ballwars;
 
 public class Player {
     public int id;
-    public float pang;
+    public double pang;
     public String isim;
     public Ball top;
 
-    public Player(int id, float pang, String isim, Ball top) {
+    public Player(int id, double pang, String isim, Ball top) {
         this.id = id;
         this.pang = pang
         this.isim = isim;
